Move Core.Pipeline to Erased.Ann.Convert
