+++ library/Backends/Michelson/test/Test/Golden.hs
@@ -5,12 +5,10 @@ module Test.Golden where
 
 import qualified Juvix.Backends.Michelson as Michelson
 import qualified Juvix.Core.Erased.Ann.Types as ErasedAnn
-import qualified Juvix.Core.Pipeline as Core
 import Juvix.Library
 import Juvix.Library.Test.Golden
 import qualified Juvix.Pipeline as Pipeline
 import Test.Tasty
-import Text.Pretty.Simple (pShowNoColor)
 
 juvixRootPath :: FilePath
 juvixRootPath = "../../../"
